+++ packages/query/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.1] - 2024-02-29
+### Fixed
+- Relation filters not taking into account block height when historical is enabled (#2276)
+
 ## [2.9.0] - 2024-01-25
 ### Added
 - gzip compression to queries (#2195)
@@ -279,7 +283,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - overwrite plugin to fix one to one unique key check
 - update query publish and docker build process
 
-[Unreleased]: https://github.com/subquery/subql/compare/query/2.9.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/query/2.9.1...HEAD
+[2.9.1]: https://github.com/subquery/subql/compare/query/2.9.0...query/2.9.1
 [2.9.0]: https://github.com/subquery/subql/compare/query/2.8.0...query/2.9.0
 [2.8.0]: https://github.com/subquery/subql/compare/query/2.7.0...query/2.8.0
 [2.7.0]: https://github.com/subquery/subql/compare/query/2.6.0...query2.7.0